Backend commit #1 1) Скорректирована БД (добавлена таблица бонусов, скорректированы остальные таблицы) 2) Добавлена модель Bonus, скорректированы DTO
* изменил место хранение БД с оперативной памяти на файл (если будет не удобно - раскомментируйте прошлый url)
* в следующем коммите реализую CRUD операции для всех моделей

+++ src/main/java/ru/coolteam/earnpocketmoney/repositories/ParentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ru.coolteam.earnpocketmoney.models.Parent;
 
 @Repository
-public interface ParentRepository extends JpaRepository<Parent, Long> {
+public interface ParentRepository extends JpaRepository<Parent, Integer> {
 }
